feat: 최단경로 기본요금 구하기 기능 추가

+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -28,7 +28,7 @@ public class PathService {
         Station target = stationService.findById(pathRequest.getTarget());
         PathFinder pathFinder = new PathFinder(lineRepository.findAll());
         StationResponses stationResponses = new StationResponses(pathFinder.findPathStations(source, target));
-        BigDecimal shortestPathLength = pathFinder.findPathLength(source, target);
-        return new PathResponse(stationResponses.getList(), shortestPathLength.doubleValue());
+        int shortestPathLength = pathFinder.findPathLength(source, target);
+        return new PathResponse(stationResponses.getList(), shortestPathLength);
     }
 }
